@ekohl 2020-11-09
The implied question was: is this a good path forward? If so, I can take a look at finishing them up.