@ghoneycutt 2020-11-09
Great work @ekohl !!

`Move facterversion_obj declaration out of the loop`

I read this as a safe way to improve existing code. +1

`Do not query for the exact facter version`

I'm understanding this as we can currently test for a specific version of Facter and after this patch, we cannot. Very likely I'm totally misunderstanding this, though if that is the case, I would not want to do that or would want an option where you can bypass it and specifically test a given version of Facter.

`Collect facts iteratively`

I read this as another optimization so +1. I'm not sure what the ordering that potentially causing test failures means and what the implications are for that.