@ekohl 2020-11-09
> I'm understanding this as we can currently test for a specific version of Facter and after this patch, we cannot. Very likely I'm totally misunderstanding this, though if that is the case, I would not want to do that or would want an option where you can bypass it and specifically test a given version of Facter.

It should still work like that (since `Gem::Version.new(facts[:facterversion]) <= Gem::Version.new(facterversion)` should also yield an exact match). However, when I ran the tests locally they failed. For example:

https://github.com/mcanevet/rspec-puppet-facts/blob/c2987a1b6cd7a9ab104aed8aa7ff61e28ce29cce/spec/rspec_puppet_facts_spec.rb#L738-L753

The queried facterversion is 3.1.2 but it expected 3.1.6 to be returned. After the second patch it returns a set of facts with version 3.0.2 since 3.1.6 > 3.1.2. I wonder if this is silently converted to a float in the filterspec by JGrep.

Given there was an explicit test for it, that makes me think it's a regression.