@ghoneycutt 2020-11-09
If you specify 3.1.2, I would expect it to respond with the same 3.1.2 or an error.