@ekohl 2020-11-09
You get the error if you specify strict versions, but that's not the default behavior.