@sanfrancrisko 2020-11-30
Just so I'm following correctly @ekohl , do you think the test expectations (that you referred to [here](https://github.com/mcanevet/rspec-puppet-facts/pull/124#issuecomment-724230473) need updated given the change in behaviour change from this PR?