@ekohl 2023-07-05
I've decided to split up the patches for a better changelog (and easier discussion): https://github.com/voxpupuli/rspec-puppet-facts/pull/151 & https://github.com/voxpupuli/rspec-puppet-facts/pull/152